Fix ImportError after refactor in autoscaler
After d761ca6606c78d4758a51502e16d65a07c14eef0 I missed some important
imports!
